01-llm-eval-harness/requirements.txt
openai
pydantic
typer
rich
python-dotenv
